Split unix-specific function into UnixFileDescription
