+++ src/shims/files.rs
@@ -7,14 +7,8 @@ use std::rc::{Rc, Weak};
 
 use rustc_abi::Size;
 
+use crate::shims::unix::UnixFileDescription;
 use crate::*;
-
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-pub(crate) enum FlockOp {
-    SharedLock { nonblocking: bool },
-    ExclusiveLock { nonblocking: bool },
-    Unlock,
-}
 
 /// Represents an open file description.
 pub trait FileDescription: std::fmt::Debug + Any {
@@ -50,37 +44,6 @@ pub trait FileDescription: std::fmt::Debug + Any {
         throw_unsup_format!("cannot write to {}", self.name());
     }
 
-    /// Reads as much as possible into the given buffer `ptr` from a given offset.
-    /// `len` indicates how many bytes we should try to read.
-    /// `dest` is where the return value should be stored: number of bytes read, or `-1` in case of error.
-    fn pread<'tcx>(
-        &self,
-        _communicate_allowed: bool,
-        _offset: u64,
-        _ptr: Pointer,
-        _len: usize,
-        _dest: &MPlaceTy<'tcx>,
-        _ecx: &mut MiriInterpCx<'tcx>,
-    ) -> InterpResult<'tcx> {
-        throw_unsup_format!("cannot pread from {}", self.name());
-    }
-
-    /// Writes as much as possible from the given buffer `ptr` starting at a given offset.
-    /// `ptr` is the pointer to the user supplied read buffer.
-    /// `len` indicates how many bytes we should try to write.
-    /// `dest` is where the return value should be stored: number of bytes written, or `-1` in case of error.
-    fn pwrite<'tcx>(
-        &self,
-        _communicate_allowed: bool,
-        _ptr: Pointer,
-        _len: usize,
-        _offset: u64,
-        _dest: &MPlaceTy<'tcx>,
-        _ecx: &mut MiriInterpCx<'tcx>,
-    ) -> InterpResult<'tcx> {
-        throw_unsup_format!("cannot pwrite to {}", self.name());
-    }
-
     /// Seeks to the given offset (which can be relative to the beginning, end, or current position).
     /// Returns the new position from the start of the stream.
     fn seek<'tcx>(
@@ -99,25 +62,14 @@ pub trait FileDescription: std::fmt::Debug + Any {
         throw_unsup_format!("cannot close {}", self.name());
     }
 
-    fn flock<'tcx>(
-        &self,
-        _communicate_allowed: bool,
-        _op: FlockOp,
-    ) -> InterpResult<'tcx, io::Result<()>> {
-        throw_unsup_format!("cannot flock {}", self.name());
-    }
-
     fn is_tty(&self, _communicate_allowed: bool) -> bool {
         // Most FDs are not tty's and the consequence of a wrong `false` are minor,
         // so we use a default impl here.
         false
     }
 
-    /// Check the readiness of file description.
-    fn get_epoll_ready_events<'tcx>(
-        &self,
-    ) -> InterpResult<'tcx, crate::shims::unix::linux::epoll::EpollReadyEvents> {
-        throw_unsup_format!("{}: epoll does not support this file description", self.name());
+    fn as_unix(&self) -> &dyn UnixFileDescription {
+        panic!("Not a unix file descriptor: {}", self.name());
     }
 }
